pyproject.toml
[tool.poetry]
name = "frogchallenge"
version = "0.1.0"
description = ""
authors = ["Rubi Mendes Passos <rubimpassos@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.3"
djangorestframework = "^3.13.1"
pandas = "^1.4.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
pytest-django = "^4.5.2"
pytest-datadir = "^1.3.1"
pytest-unordered = "^0.4.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
